TFN tgui say cherry-pick (#104)
Fixes TGUI menu not focusing on creation (#668)

* ghrhrrhhrhr

* 90395

* 90608

* Update communication.dm

* Update communication.dm

* nghgngh

* wew

* aaa

* AAAAAA

* aaaa

* update papers

* Revert "update papers"

This reverts commit 1514aa19cc4596f94726b16dd33a658bab61fec2.

* Revert "aaaa"

This reverts commit d2f724b50e15ab8a88742c9da3281b7d9b870270.
