+++ code/datums/keybinding/communication.dm
@@ -11,7 +11,8 @@
 	. = ..()
 	if(.)
 		return
-	winset(user, null, "command=[user.tgui_say_create_open_command(SAY_CHANNEL)]")
+	winset(user, null, "command=[user.tgui_say_create_open_command(SAY_CHANNEL)];")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 
 /datum/keybinding/client/communication/radio
@@ -25,6 +26,7 @@
 	if(.)
 		return
 	winset(user, null, "command=[user.tgui_say_create_open_command(RADIO_CHANNEL)]")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 
 /datum/keybinding/client/communication/ooc
@@ -38,6 +40,7 @@
 	if(.)
 		return
 	winset(user, null, "command=[user.tgui_say_create_open_command(OOC_CHANNEL)]")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 
 /datum/keybinding/client/communication/me
@@ -51,6 +54,7 @@
 	if(.)
 		return
 	winset(user, null, "command=[user.tgui_say_create_open_command(ME_CHANNEL)]")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 
 // TFN EDIT START
@@ -65,10 +69,11 @@
 	if(.)
 		return
 	winset(user, null, "command=[user.tgui_say_create_open_command(LOOC_CHANNEL)]")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 
 /datum/keybinding/client/communication/whisper
-	hotkey_keys = list("Y")
+	hotkey_keys = list("Unbound")
 	name = WHIS_CHANNEL
 	full_name = "IC Whisper"
 	keybind_signal = COMSIG_KB_CLIENT_WHISPER_DOWN
@@ -78,5 +83,6 @@
 	if(.)
 		return
 	winset(user, null, "command=[user.tgui_say_create_open_command(WHIS_CHANNEL)]")
+	winset(user, "tgui_say.browser", "focus=true")
 	return TRUE
 // TFN EDIT END
