+++ code/modules/mob/mob_say.dm
@@ -25,7 +25,7 @@
 	flavor_text = flavor // We don't have to worry about flavor text length due to examine overflow.
 // TFN EDIT END
 ///Whisper verb
-/mob/verb/whisper_verb(message as text)
+/mob/verb/whisper_verb(message as text|null)
 	set name = "Whisper"
 	set category = "IC"
 	if(GLOB.say_disabled)	//This is here to try to identify lag problems
@@ -38,7 +38,7 @@
 	say(message, language) //only living mobs actually whisper, everything else just talks
 // TFN EDIT START
 ///The me emote verb
-/mob/verb/me_verb(message as text)
+/mob/verb/me_verb(message as text|null)
 	set name = "Me"
 	set category = "IC"
