--- tgui/packages/tgui-say/515/constants.ts
@@ -1,34 +1,0 @@
-/** Window sizes in pixels */
-export enum WINDOW_SIZES {
-  small = 30,
-  medium = 50,
-  large = 70,
-  width = 231,
-}
-
-/** Line lengths for autoexpand */
-export enum LINE_LENGTHS {
-  small = 22,
-  medium = 45,
-}
-
-/**
- * Radio prefixes.
- * Displays the name in the left button, tags a css class.
- */
-export const RADIO_PREFIXES = {
-  ':a ': 'Hive',
-  ':b ': 'io',
-  ':c ': 'Cmd',
-  ':e ': 'Engi',
-  ':g ': 'Cling',
-  ':m ': 'Med',
-  ':n ': 'Sci',
-  ':o ': 'AI',
-  ':p ': 'Ent',
-  ':s ': 'Sec',
-  ':t ': 'Synd',
-  ':u ': 'Supp',
-  ':v ': 'Svc',
-  ':y ': 'CCom',
-} as const;
