--- tgui/packages/tgui-say/515/styles/button.scss
@@ -1,33 +1,0 @@
-@use 'sass:color';
-@use './colors.scss';
-
-.old-button {
-  align-items: center;
-  background-color: colors.$button;
-  border-radius: 0.3rem;
-  border: thin solid;
-  display: flex;
-  flex-grow: 1;
-  font-family: inherit;
-  font-size: 0.9rem;
-  font-weight: bold;
-  justify-content: center;
-  padding: 0;
-  width: 2.6rem;
-  &:hover {
-    background-color: color.adjust(
-      colors.$button,
-      $lightness: 10%,
-      $space: hsl
-    );
-  }
-}
-
-.old-button-lightMode {
-  background-color: colors.$lightBorder;
-  border: none;
-  color: black;
-  &:hover {
-    background-color: colors.$lightHover;
-  }
-}
