--- tgui/packages/tgui-say/515/styles/content.scss
@@ -1,14 +1,0 @@
-.old-center {
-  display: flex;
-  flex: 1 1 0;
-  height: 100%;
-  width: 100%;
-}
-
-.old-input {
-  display: flex;
-  flex: 1 1 0;
-  height: 100%;
-  width: 100%;
-  font-family: 'Consolas', monospace;
-}
