--- tgui/packages/tgui-say/515/styles/dragzone.scss
@@ -1,39 +1,0 @@
-@use 'sass:color';
-@use './colors.scss';
-
-$dragSize: 0.6rem;
-$borderSize: 0.2rem;
-
-.old-dragzone-horizontal {
-  border-left: $borderSize solid;
-  border-right: $borderSize solid;
-  color: transparent;
-  width: 100%;
-  height: $dragSize;
-}
-.old-dragzone-left {
-  border-left: $borderSize solid;
-}
-
-.old-dragzone-right {
-  border-right: $borderSize solid;
-}
-
-.old-dragzone-vertical {
-  color: transparent;
-  height: 100%;
-  width: $dragSize;
-}
-
-.old-dragzone-top {
-  border-top: $borderSize solid;
-}
-
-.old-dragzone-bottom {
-  border-bottom: $borderSize solid;
-}
-
-/** Lightmode static theme */
-.old-dragzone-lightMode {
-  border-color: colors.$lightBorder;
-}
