--- tgui/packages/tgui-say/515/styles/main.scss
@@ -1,70 +1,0 @@
-@use 'sass:meta';
-@use 'sass:color';
-@use './colors.scss';
-
-// Core styles
-@include meta.load-css('~tgui/styles/reset.scss');
-// Atomic styles
-@include meta.load-css('~tgui/styles/atomic/text.scss');
-// External styles
-@include meta.load-css('~tgui-core/styles/components/TextArea');
-// Local styles
-@include meta.load-css('./button.scss');
-@include meta.load-css('./content.scss');
-@include meta.load-css('./dragzone.scss');
-@include meta.load-css('./textarea.scss');
-@include meta.load-css('./window.scss');
-
-@keyframes gradient {
-  0% {
-    background-position: 0 0;
-  }
-
-  100% {
-    background-position: 100% 0;
-  }
-}
-
-@each $channel, $color in colors.$channel-map {
-  $darkened: color.adjust($color, $lightness: -20%, $space: hsl);
-
-  .old-button-#{$channel} {
-    border-color: color.adjust($color, $lightness: -10%, $space: hsl);
-    color: $color;
-    &:hover {
-      border-color: color.adjust($color, $lightness: 10%, $space: hsl);
-      color: color.adjust($color, $lightness: 5%, $space: hsl);
-    }
-  }
-
-  .old-dragzone-#{$channel} {
-    border-color: $darkened;
-  }
-
-  .old-textarea-#{$channel} {
-    color: $color;
-  }
-
-  .old-window-#{$channel} {
-    &:after {
-      animation: gradient 10s linear infinite;
-      background: linear-gradient(
-        to right,
-        color.adjust($color, $lightness: -35%, $space: hsl),
-        $color,
-        color.adjust($color, $lightness: 10%, $space: hsl),
-        $color,
-        color.adjust($color, $lightness: -35%, $space: hsl)
-      );
-      background-position: 0% 0%;
-      background-size: 500% auto;
-      bottom: 0px;
-      content: '';
-      height: 2px;
-      left: 0px;
-      position: absolute;
-      right: 0px;
-      z-index: 999;
-    }
-  }
-}
