--- tgui/packages/tgui-say/515/styles/textarea.scss
@@ -1,11 +1,0 @@
-.old-textarea {
-  align-items: center;
-  background: transparent;
-  border: none;
-  display: flex;
-  flex-grow: 4;
-  font-family: inherit;
-  font-size: 1.1rem;
-  overflow: hidden;
-  margin: 0.1rem 0 0 0.4rem;
-}
