--- tgui/packages/tgui-say/516/constants.ts
@@ -1,34 +1,0 @@
-/** Window sizes in pixels */
-export enum WindowSize {
-  Small = 30,
-  Medium = 50,
-  Large = 70,
-  Width = 231,
-}
-
-/** Line lengths for autoexpand */
-export enum LineLength {
-  Small = 20,
-  Medium = 40,
-}
-
-/**
- * Radio prefixes.
- * Displays the name in the left button, tags a css class.
- */
-export const RADIO_PREFIXES = {
-  ':a ': 'Hive',
-  ':b ': 'io',
-  ':c ': 'Cmd',
-  ':e ': 'Engi',
-  ':g ': 'Cling',
-  ':m ': 'Med',
-  ':n ': 'Sci',
-  ':o ': 'AI',
-  ':p ': 'Ent',
-  ':s ': 'Sec',
-  ':t ': 'Synd',
-  ':u ': 'Supp',
-  ':v ': 'Svc',
-  ':y ': 'CCom',
-} as const;
