+++ tgui/packages/tgui-say/index.tsx
@@ -1,8 +1,6 @@
 import { createRoot, Root } from 'react-dom/client';
 
-// TODO: remove this once we're on 516
-import { TguiSay } from './515/TguiSay';
-import { TguiSay as NewSay } from './516/TguiSay';
+import { TguiSay } from './TguiSay';
 
 let reactRoot: Root | null = null;
 
@@ -14,9 +12,5 @@ document.onreadystatechange = function () {
     reactRoot = createRoot(root!);
   }
 
-  if (Byond.BLINK) {
-    reactRoot.render(<NewSay />);
-  } else {
-    reactRoot.render(<TguiSay />);
-  }
+  reactRoot.render(<TguiSay />);
 };
