+++ tgui/packages/tgui-say/styles/main.scss
@@ -2,13 +2,7 @@
 @use 'sass:color';
 @use './colors.scss';
 
-// Core styles
 @include meta.load-css('~tgui/styles/reset');
-// Atomic styles
-@include meta.load-css('~tgui/styles/atomic/text.scss');
-// External styles
-@include meta.load-css('~tgui-core/styles/components/TextArea');
-// Local styles
 @include meta.load-css('./styles.scss');
 
 @each $channel, $color in colors.$channel-map {
@@ -17,6 +11,7 @@
   .button-#{$channel} {
     border-color: color.adjust($color, $lightness: -10%, $space: hsl);
     color: $color;
+    user-select: none;
     &:hover {
       border-color: color.adjust($color, $lightness: 10%, $space: hsl);
       color: color.adjust($color, $lightness: 5%, $space: hsl);
