+++ tgui/packages/tgui-say/styles/styles.scss
@@ -31,6 +31,7 @@
 
 .window-lightMode {
   background-color: colors.$lightMode;
+  border: 2px solid colors.$lightBorder;
 }
 
 /** Window sizes */
@@ -50,7 +51,7 @@
   background-color: black;
   display: grid;
   font: 'tgfont';
-  grid-template-columns: 3.5rem 1fr;
+  grid-template-columns: 4rem 1fr;
   inset: 2px;
   overflow: hidden;
   padding: 2px;
@@ -70,11 +71,10 @@
   font-weight: bold;
   outline: none;
   overflow: hidden;
-  padding: 0.1rem 0.2rem;
+  padding: 0.1rem;
   text-align: center;
   text-overflow: ellipsis;
   white-space: nowrap;
-  width: 100%;
   &:hover {
     background-color: color.adjust(
       colors.$button,
@@ -101,6 +101,7 @@
   margin: 0.1rem 0 0 0.4rem;
   outline: none;
   resize: none;
+  overflow: hidden;
   &::-webkit-scrollbar {
     width: 0.8rem;
   }
@@ -116,4 +117,8 @@
       $lightness: 10% * colors.$scrollbar-color-multiplier
     );
   }
+}
+
+.textarea-large {
+  overflow-y: scroll;
 }
